move nodder, emoter, blinker to their own files
