+++ avatars/Emoter.js
@@ -1,0 +1,9 @@
+export default class Emoter {
+  constructor() {
+
+  }
+
+  update() {
+
+  }
+}
