+++ avatars/Nodder.js
@@ -1,0 +1,9 @@
+export default class Nodder {
+  constructor() {
+
+  }
+
+  update() {
+
+  }
+}
