+++ avatars/avatars.js
@@ -47,6 +47,9 @@ import {
 } from './util.mjs';
 import metaversefile from 'metaversefile';
 import { animationMappingConfig } from './AnimationMapping.js';
+import Emoter from './Emoter.js'
+import Blinker from './Blinker.js'
+import Nodder from './Nodder.js'
 import Looker from './Looker.js'
 
 const textEncoder = new TextEncoder();
@@ -835,78 +838,14 @@ const _clearXZ = (dst, isPosition) => {
   }
 };
 
-class Blinker {
-  constructor() {
-    this.mode = 'ready';
-    this.waitTime = 0;
-    this.lastTimestamp = 0;
-  }
-  update(now) {
-    const _setOpen = () => {
-      this.mode = 'open';
-      this.waitTime = (0.5 + 0.5 * Math.random()) * 3000;
-      this.lastTimestamp = now;
-    };
 
-    switch (this.mode) {
-      case 'ready': {
-        _setOpen();
-        return 0;
-      }
-      case 'open': {
-        const timeDiff = now - this.lastTimestamp;
-        if (timeDiff > this.waitTime) {
-          this.mode = 'closing';
-          this.waitTime = 100;
-          this.lastTimestamp = now;
-        }
-        return 0;
-      }
-      case 'closing': {
-        const f = Math.min(Math.max((now - this.lastTimestamp) / this.waitTime, 0), 1);
-        if (f < 1) {
-          return f;
-        } else {
-          this.mode = 'opening';
-          this.waitTime = 100;
-          this.lastTimestamp = now;
-          return 1;
-        }
-      }
-      case 'opening': {
-        const f = Math.min(Math.max((now - this.lastTimestamp) / this.waitTime, 0), 1);
-        if (f < 1) {
-          return 1 - f;
-        } else {
-          _setOpen();
-          return 0;
-        }
-      }
-    }
-  }
-}
-class Nodder {
-  constructor() {
 
-  }
-  update() {
-
-  }
-}
 // const g = new THREE.BoxBufferGeometry(0.05, 0.05, 0.05);
 // const m = new THREE.MeshBasicMaterial({ color: 0xFF00FF });
 // const testMesh = new THREE.Mesh(g, m);
 // scene.add(testMesh);
 
 
-class Emoter {
-  constructor() {
-
-  }
-  update() {
-
-  }
-}
 
 class Avatar {
 	constructor(object, options = {}) {
